requirements.txt
streamlit>=1.28.0
pandas>=2.0.0
numpy>=1.24.0
scikit-learn>=1.3.0
sentence-transformers>=2.2.2
PyPDF2>=3.0.1
python-docx>=0.8.11
spacy>=3.5.0
pymongo>=4.0.0
